% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_hyp.R
\name{CTDC}
\alias{CTDC}
\title{Composition descriptor (CTDC)}
\usage{
CTDC(x)
}
\arguments{
\item{x}{A vector of k-mers (character vector)}
}
\value{
A matrix with the Composition descriptor (dim: 21)
}
\description{
This function calculates the Composition descriptor
}
\note{
This is an internal function used inside predict_hyp
}
\references{
Inna Dubchak, Ilya Muchink, Stephen R. Holbrook and Sung-Hou Kim. Prediction of protein folding class using global description of amino acid sequence. Proceedings of the National Academy of Sciences. USA, 1995, 92, 8700-8704.
}
\seealso{
\code{\link[ragp]{predict_hyp} \link[protr]{extractCTDC}}
}
