% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_hyp.R
\name{extractMBdesc}
\alias{extractMBdesc}
\title{Normalized Moreau-Broto Autocorrelation Descriptor}
\usage{
extractMBdesc(x, mat, nlag)
}
\arguments{
\item{x}{A vector of k-mers}

\item{mat}{A matrix with amino acid atributes. Columns are amino acids, rows are atributes}

\item{nlag}{The lag parameter}
}
\value{
A matrix with the Normalized Moreau-Broto Autocorrelation Descriptor (dim: nlag * nrow(mat))
}
\description{
This function calculates the normalized Moreau-Broto autocorrelation descriptor.
}
\details{
Normalized Moreau-Broto Autocorrelation Descriptor is calculated based on the following amino acid attributes:
CIDH920105 - Normalized average hydrophobicity scales, BHAR880101 - Average flexibility indices,
CHAM820102 - Free energy of solution in water, kcal/mole, BIGC670101 - Residue volume,
CHAM810101 - Polarizability parameter and DAYM780201 - Relative mutability
}
\note{
This is an internal function used inside predict_hyp
}
\seealso{
\code{\link[ragp]{predict_hyp} \link[protr]{extractMoreauBroto}}
}
\author{
original R code by Nan Xiao, modified by Milan Dragićević
}
