% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_big_pi.R
\name{get_big_pi}
\alias{get_big_pi}
\title{Scraping big-PI Plant Predictor web server.}
\source{
\url{http://mendel.imp.ac.at/gpi/plant_server.html}
}
\usage{
get_big_pi(sequence, id, simplify = TRUE, sleep = NULL, verbose = TRUE)
}
\arguments{
\item{sequence}{A vector of strings representing protein amino acid sequences}

\item{id}{A vector of strings representing protein identifiers}

\item{simplify}{A bolean indicating the type of returned object, defaults to TRUE}

\item{sleep}{A numeric indicating the pause in seconds betwean server calls, at default set to 1}

\item{verbose}{Bolean wheather to print out the total score for each sequence, defaults to TRUE}
}
\value{
If simplify == TRUE:
A data frame with columns:
\describe{
  \item{omega_site}{Character, indicating the sequence position of the highest scoring omega-site}
  \item{Quality}{Character, indicating the quality of the highest scoring omega-site}
  \item{PValue}{Numeric, indicating the p-value for the prediction of the highest scoring omega-site"}
  \item{id}{Character, indicating the protein identifier}
  }

If simplify == FALSE:
A list of predictions, each element named according to the sequence id, containg a two element list:
\describe{
  \item{prediction}{data frame, resembling the one returned by simplify == TRUE, along with alternative site predictions (if present)}
  \item{calculation}{data frame, with profile dependant and profile independant scores}
  }
}
\description{
big-PI Plant Predictor is a web server utlizing a scoring algorithm for prediction of GPI modification sites in plants.
}
\examples{
library(ragp)
data(at_nsp)

#indexes of some sequences in at_nsp
ind <- c(129, 145, 147, 160, 170)

big_pi_pred <- get_big_pi(sequence = at_nsp$sequence[ind],
                          id = at_nsp$Transcript.id[ind],
                          simplify = FALSE)
                          
big_pi_pred <- get_big_pi(sequence = at_nsp$sequence[ind],
                          id = at_nsp$Transcript.id[ind],
                          simplify = TRUE)
}
\references{
Eisenhaber B. Wildpaner M. Schultz CJ. Borner GHH. Dupree P. Eisenhaber F. (2003) Glycosylphosphatidylinositol lipid anchoring of plant proteins. Sensitive prediction from sequence- and genome-wide studies for Arabidopsis and rice. Plant Physiology 133(4): 1691-701
}
