% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdd.R
\name{get_cdd}
\alias{get_cdd}
\alias{get_cdd.default}
\alias{get_cdd.character}
\alias{get_cdd.data.frame}
\alias{get_cdd.list}
\alias{get_cdd.AAStringSet}
\title{Query the Conserved Domain Database.}
\source{
\url{https://www.ncbi.nlm.nih.gov/Structure/bwrpsb/bwrpsb.cgi}
}
\usage{
get_cdd(data, ...)

\method{get_cdd}{character}(
  data,
  output = c("short", "standard", "full"),
  splitter = 100L,
  progress = FALSE,
  evalue = NULL,
  bitscore = NULL,
  ...
)

\method{get_cdd}{data.frame}(data, sequence, id, ...)

\method{get_cdd}{list}(data, ...)

\method{get_cdd}{default}(data = NULL, sequence, id, ...)

\method{get_cdd}{AAStringSet}(data, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class \code{\link[seqinr]{SeqFastaAA}} resulting from \code{\link[seqinr]{read.fasta}} call. Alternatively an \code{\link[Biostrings]{AAStringSet}} object. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{output}{One of "short", "standard" or "full". Default is "short". "Short" returns the highest scoring hit, for each region of the query sequence; "standard" returns the best-scoring hit from each source database, for each region of the query sequence; "full" returns the complete set of hits.}

\item{splitter}{An integer indicating the number of sequences to be in each .fasta file that is to be sent to the server. Default is 100. Change only in case of a server side error. Accepted values are in range of 1 to 4000.}

\item{progress}{Boolean, whether to show the progress bar, at default set to FALSE.}

\item{evalue}{Numeric, all sequences with E-value lower or equal to this value will be retained in the function output. Used to filter out low similarity matches. If set some queried sequences might be discarded from the output. Suggested values: 1e-2 - 1e-5.}

\item{bitscore}{Numeric, all sequences with bitscore greater or equal to this value will be retained in the function output. Used to filter out low similarity. If set some queried sequences might be discarded from the output. Suggested values: 10 - 20.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}
}
\value{
A data frame with columns: 
\describe{
  \item{id}{Character, as from input}
  \item{hit.type}{Character, CD-Search results can include hit types that represent various confidence levels (specific hits, non-specific hits) and domain model scope (superfamilies, multi-domains).}
  \item{pssm.id}{Character, unique identifier for a domain models position-specific scoring matrix (PSSM).}
  \item{align_start}{Integer, start of the alignement in the query protein sequence.}
  \item{align_end}{Integer, end of the alignement in the query protein sequence.}
  \item{evalue}{Numeric, the expect value, or E-value, indicates the statistical significance of the hit as the likelihood the hit was found by chance.}
  \item{bitscore}{Numeric, this values is derived from the raw alignment score in which the statistical properties of the scoring system used have been taken into account.}
  \item{acc}{Character, the accession number of the hit, which can either be a domain model or a superfamily cluster.}
  \item{desc}{Character, the short name of a conserved domain, which concisely defines the domain.}  
  \item{incomplete}{Character, if the hit to a conserved domain is partial (i.e., if the alignment found by RPS-BLAST omitted more than 20 percent of the CDs extent at either the n- or c-terminus or both), this column will be populated with one of the following values: N - incomplete at the N-terminus, C - incomplete at the C-terminus, NC - incomplete at both the N-terminus and C-terminus. If the hit to a conserved domain is complete, then this column will be populated with a dash (-).} 
  \item{superfamily}{Character, the short name of a conserved domain, which concisely defines the domain.}
  \item{definition}{Character, this column is populated only for domain models that are specific or non-specific hits, and it lists the accession number of the superfamily to which the domain model belongs.}   
  }
}
\description{
Conserved Domain Database is a protein annotation resource that consists of a collection of well-annotated multiple sequence alignment models for ancient domains and full-length proteins.
}
\note{
This function creates temporary files in the working directory.
}
\examples{

library(ragp)
cdd_pred <- get_cdd(data = at_nsp[1:10,],
                    sequence,
                    Transcript.id)
                    
cdd_pred[,1:11]

}
\references{
Lu S, Wang J, Chitsaz F, Derbyshire MK, Geer RC, Gonzales NR, Gwadz M, Hurwitz DI, Marchler GH, Song JS, Thanki N, Yamashita RA, Yang M, Zhang D, Zheng C, Lanczycki CJ, Marchler-Bauer A (2020) CDD/SPARCLE: the conserved domain database in 2020, Nucleic Acids Res. 48(D1):D265-D268. doi: https://doi.org/10.1093/nar/gkz991
}
\seealso{
\code{\link[ragp]{get_hmm}} \code{\link[ragp]{plot_prot}}
}
