% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hmm.R
\name{get_hmm}
\alias{get_hmm}
\title{Scraping hmmer web server.}
\source{
\url{https://www.ebi.ac.uk/Tools/hmmer/search/hmmscan}
}
\usage{
get_hmm(data = NULL, sequence, id, verbose = TRUE, sleep = 1)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from seqinr::read.fasta call.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{verbose}{Bolean, whether to print out the output for each sequence, defaults to T}

\item{sleep}{Numeric indicating the pause in seconds between server calls, at default set to 1}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, as supplied in the function call}
  \item{name}{Character, PFAM family name}
  \item{acc}{Character, PFAM family accession}
  \item{desc}{Character, PFAM family description}
  \item{clan}{Character, PFAM clan}
  \item{align_start}{Numeric, start of domain alignment in query sequence}
  \item{align_end}{Numeric, end of domain alignment in query sequence}
  \item{model_start}{Numeric, start of alignment in domain model}
  \item{model_end}{Numeric, end of alignment in domain model}
  \item{ievalue}{Numeric, the "independent E-value", the E-value that the sequence/profile comparison would have received if this were the only domain envelope found in it, excluding any others. This is a stringent measure of how reliable this particular domain may be. The independent E-value uses the total number of targets in the target database.}
  \item{cevalue}{Numeric, the "conditional E-value", a permissive measure of how reliable this particular domain may be.}
  \item{bitscore}{Numeric, the domain bit score.}
  \item{reported}{Logical, is the result reported on the hmmer site. The hmmer web server outputs more hmm profile matches than it presents to the user. Results below a certain threshold are not reported (hidden) on the site.}
}
}
\description{
hmmer web server offers biosequence analysis using profile hidden Markov Models. This function allows searching
of a protein sequence vs a profile-HMM database (Pfam-A).
}
\examples{

pfam_pred <- get_hmm(sequence = at_nsp$sequence[1],
                    id = at_nsp$Transcript.id[1])

}
\seealso{
\code{\link[ragp]{pfam2go}}
}
