% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_netGPI.R
\name{get_netGPI}
\alias{get_netGPI}
\alias{get_netGPI.default}
\alias{get_netGPI.character}
\alias{get_netGPI.data.frame}
\alias{get_netGPI.list}
\title{Query NetGPI - 1.1 web server.}
\source{
\url{https://services.healthtech.dtu.dk/service.php?NetGPI-1.1}
}
\usage{
get_netGPI(data, ...)

\method{get_netGPI}{character}(data, splitter = 2500L, attempts = 2, progress = FALSE, ...)

\method{get_netGPI}{data.frame}(data, sequence, id, ...)

\method{get_netGPI}{list}(data, ...)

\method{get_netGPI}{default}(data = NULL, sequence, id, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from \code{\link[seqinr]{read.fasta}} call. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{splitter}{An integer indicating the number of sequences to be in each .fasta file that is to be sent to the server. Defaults to 2500. Change only in case of a server side error. Accepted values are in range of 1 to 5000.}

\item{attempts}{Integer, number of attempts if server unresponsive, at default set to 2.}

\item{progress}{Boolean, whether to show the progress bar, at default set to FALSE.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, as from input}
  \item{length}{Integer, length of the protein sequence}
  \item{is.gpi}{Logical, is the protein predicted to be GPI anchored.}
  \item{omega_site}{Integer, indicating the sequence position of the omega-site.}
  \item{likelihood}{Numeric, likelihood of the prediction.}}
}
\description{
NetGPI server offers GPI Anchor predictions
}
\note{
This function creates temporary files in the working directory.
}
\examples{

library(ragp)
netGPI_pred <- get_netGPI(data = at_nsp[1:10,],
                          sequence,
                          Transcript.id)
netGPI_pred

}
\references{
Gislason MH. Nielsen H. Armenteros JA. AR Johansen AR. (2019) Prediction of GPI-Anchored proteins with pointer neural networks. bioRxiv. doi: https://doi.org/10.1101/838680
}
\seealso{
\code{\link[ragp]{get_big_pi}} \code{\link[ragp]{get_pred_gpi}}
}
