% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phobius_file.R
\name{get_phobius_file}
\alias{get_phobius_file}
\title{Scraping Phobius web server.}
\source{
\url{http://phobius.sbc.su.se/}
}
\usage{
get_phobius_file(file)
}
\arguments{
\item{file}{path to the fasta formatted file to be analyzed}
}
\value{
A data frame with columns:
\describe{
\item{Name}{Character, name of the submitted sequence.}
\item{tm}{Integer, the number of predicted transmembrane segments.}
\item{sp}{Character, Y/0 indicator if a signal peptide was predicted or not.}
\item{prediction}{Character string, predicted topology of the protein.}
\item{cut_site}{Integer, first amino acid after removal of the signal peptide}
\item{is.phobius}{Logical, did Phobius predict the presence of a signal peptide}
}
}
\description{
Phobius web server is a combined transmembrane topology and signal peptide (N-sp) predictor. Currently only "normal prediction" of signal peptides is supported by the function.
}
\details{
The topology (prediction column of the result) is given as the position of the transmembrane helices separated by 'i' if the loop is on the cytoplasmic or 'o' if it is on the non-cytoplasmic side. A signal peptide is given by the position of its h-region separated by a n and a c, and the position of the last amino acid in the signal peptide and the first of the mature protein separated by a /.
}
\examples{

data(at_nsp)

#create a fasta file:
library(seqinr)
write.fasta(sequence = strsplit(at_nsp$sequence[1:20], ""),
            name = at_nsp$Transcript.id[1:20],
            file = "at_nsp.fasta")

#get predictions
phobius_pred <- get_phobius_file("at_nsp.fasta")
}
\references{
Kall O. Krogh A. Sonnhammer E. L. L. (2004) A Combined Transmembrane Topology and Signal Peptide Prediction Method. Journal of Molecular Biology 338(5): 1027-1036
}
\seealso{
\code{\link[ragp]{get_signalp_file}}
}
