% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targetp.R
\name{get_targetp}
\alias{get_targetp}
\alias{get_targetp.default}
\alias{get_targetp.character}
\alias{get_targetp.data.frame}
\alias{get_targetp.list}
\title{Query TargetP web server.}
\source{
\url{http://www.cbs.dtu.dk/services/TargetP/}
}
\usage{
get_targetp(data, ...)

\method{get_targetp}{character}(data, org_type = c("non_plant", "plant"),
  cutoffs = c("winner_takes_all", "spec95", "spec90", "custom"),
  tcut = NULL, pcut = NULL, scut = NULL, ocut = NULL,
  splitter = 200, sleep = 3, attempts = 2, progress = FALSE, ...)

\method{get_targetp}{data.frame}(data, sequence, id, ...)

\method{get_targetp}{list}(data, ...)

\method{get_targetp}{default}(data = NULL, sequence, id, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from \code{\link[seqinr]{read.fasta}} call. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{org_type}{One of c("non_plant", "plant"), defaults to "plant". Which models should be used for prediction.}

\item{cutoffs}{One of c("winner_takes_all", "spec95", "spec90", "custom"), defaults to "winner_takes_all". If "winner_takes_all" no cutoffs are specified, if "spec95" or "spec90" are selected, predefined set of cutoffs that yielded >0.95 or >0.9 specificity on the TargetP test sets. If "custom" specified user defined cutoffs should be specified in "tcut", "pcut", "scut", ocut".}

\item{tcut}{A numeric value, with range 0 - 1, defaults to 0 (cutoff = "winner_takes_all"). mTP user specified cutoff.}

\item{pcut}{A numeric value, with range 0 - 1, defaults to 0 (cutoff = "winner_takes_all"). cTP user specified cutoff.}

\item{scut}{A numeric value, with range 0 - 1, defaults to 0 (cutoff = "winner_takes_all"). SP user specified cutoff.}

\item{ocut}{A numeric value, with range 0 - 1, defaults to 0 (cutoff = "winner_takes_all"). User specified cutoff for "other" (not with mTP, cTP, SP).}

\item{splitter}{An integer indicating the number of sequences to be in each .fasta file that is to be sent to the server. Defaults to 200. Change only in case of a server side error. Accepted values are in range of 1 to 2000.}

\item{sleep}{A numeric indicating the pause in seconds between server calls, at default set to 1}

\item{attempts}{Integer, number of attempts if server unresponsive, at default set to 2.}

\item{progress}{Boolean, whether to show the progress bar, at default set to FALSE.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}
}
\value{
A data frame with columns:
\describe{
  \item{Name}{Character, sequence name truncated to 20 characters}
  \item{Len}{Integer, length of analyzed sequence}
  \item{cTP}{Numeric, final NN sequence score to contain a chloroplast transit peptide (cTP). ChloroP is used to predict cleavage sites for cTP}
  \item{mTP}{Numeric, final NN sequence score to contain a mitochondrial targeting peptide (mTP).}
  \item{SP}{Numeric, final NN sequence score to contain a secretory pathway signal peptide (SP). SignalP is used to predict cleavage sites for SP}
  \item{other}{Numeric, final NN sequence score of a sequence not to contain mTP, SP or cTP}
  \item{Loc}{Character, one of C (chloroplast), M (mitochondrion), S (secretory pathway), - (any other location) or * ("don't know"; indicates that cutoff restrictions were set and the winning network output score was below the requested cutoff for that category.) }
  \item{TPlen}{Integer, predicted presequence length}
  \item{RC}{Integer, reliability class, from 1 to 5, where 1 indicates the strongest prediction. RC is a measure of the size of the difference ('diff') between the highest (winning) and the second highest output scores. There are 5 reliability classes, defined as follows: 1 : diff > 0.800, 2 : 0.800 > diff > 0.600, 3 : 0.600 > diff > 0.400, 4 : 0.400 > diff > 0.200, 5 : 0.200 > diff. Thus, the lower the value of RC the safer the prediction.}
  \item{is.targetp}{Logical, did TargetP predict the presence of a signal peptide}
  }
}
\description{
TargetP 1.1 predicts the subcellular location of eukaryotic proteins. The location assignment is based on the predicted presence of any of the N-terminal presequences: chloroplast transit peptide (cTP), mitochondrial targeting peptide (mTP) or secretory pathway signal peptide (SP). TargetP uses ChloroP and SignalP to predict cleavage sites for cTP and SP, respectively. For the sequences predicted to contain an N-terminal presequence a potential cleavage site is also predicted.
}
\note{
This function creates temporary files in the working directory. Protein ids should be shorter then 20 characters due to server side truncation.
}
\examples{
library(ragp)
data(at_nsp)
targetp_pred <- get_targetp(at_nsp[1:20,],
                            sequence,
                            Transcript.id)
targetp_pred     
                      
}
\references{
Emanuelsson O, Nielsen H, Brunak S,von Heijne G. (2000) Predicting subcellular localization of proteins based on their N-terminal amino acid sequence. J. Mol. Biol.300: 1005-1016
}
\seealso{
\code{\link[ragp]{get_signalp}}
}
