% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmhmm.R
\name{get_tmhmm}
\alias{get_tmhmm}
\alias{get_tmhmm.default}
\alias{get_tmhmm.character}
\alias{get_tmhmm.data.frame}
\alias{get_tmhmm.list}
\alias{get_tmhmm.AAStringSet}
\title{Query TMHMM-2.0 web server.}
\source{
\url{https://services.healthtech.dtu.dk/service.php?TMHMM-2.0}
}
\usage{
get_tmhmm(data, ...)

\method{get_tmhmm}{character}(data, splitter = 2500L, attempts = 2, progress = FALSE, ...)

\method{get_tmhmm}{data.frame}(data, sequence, id, ...)

\method{get_tmhmm}{list}(data, ...)

\method{get_tmhmm}{default}(data = NULL, sequence, id, ...)

\method{get_tmhmm}{AAStringSet}(data, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from \code{\link[seqinr]{read.fasta}} call. Alternatively an `AAStringSet` object. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{splitter}{An integer indicating the number of sequences to be in each .fasta file that is to be sent to the server. Default is 2500. Change only in case of a server side error. Accepted values are in range of 1 to 10000.}

\item{attempts}{Integer, number of attempts if server unresponsive, at default set to 2.}

\item{progress}{Boolean, whether to show messages of the job id for each batch. Default is FALSE}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, name of the submitted sequence.}
  \item{length}{Integer, length of the protein sequence}
  \item{ExpAA}{Numeric, the expected number of amino acids in transmembrane helices.}
  \item{First60}{Numeric, the expected number of amino acids in transmembrane helices in the first 60 amino acids of the protein.}
  \item{tm}{Integer, the number of predicted transmembrane segments.}
  \item{prediction}{Character string, predicted topology of the protein.}}
}
\description{
TMHMM server offers prediction of transmembrane helices in proteins
}
\note{
This function creates temporary files in the working directory. If something goes wrong during communication with the server and progress was set to TRUE, predictions can be obtained using the web address `paste("https://services.healthtech.dtu.dk/cgi-bin/webface2.cgi?jobid=", jobid, "&wait=20", sep = "")`.
}
\examples{
library(ragp)
tmhmm_pred <- get_tmhmm(data = at_nsp[1:10,],
                        sequence,
                        Transcript.id)
tmhmm_pred

}
\references{
Krogh A, Larsson B, von Heijne G, Sonnhammer EL (2001) Predicting transmembrane protein topology with a hidden Markov model: application to complete genomes. J Mol Biol 305(3):567-80.
}
\seealso{
\code{\link[ragp]{get_phobius}}
}
