% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_hyp.R
\name{getAAindex}
\alias{getAAindex}
\title{Extract amino acid attributes}
\usage{
getAAindex(x, AA)
}
\arguments{
\item{x}{A vector of k-mers (character vector)}

\item{AA}{a matrix of amino acid attributes with attributes as rownames and amino acids as column names}
}
\value{
A matrix with specified amino acid attributes (dim: nchar(x) * ncol(AA)
}
\description{
This function extracts amino acid attributes
}
\note{
This is an internal function used inside predict_hyp
}
\seealso{
\code{\link[ragp]{predict_hyp}}
}
