% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_hyp.R
\name{getKmer}
\alias{getKmer}
\title{Extract simetric k-mers around prolines}
\usage{
getKmer(sequence, id, kmer)
}
\arguments{
\item{sequence}{protein sequence in the form of a string}

\item{id}{protein id in the form of a string}

\item{kmer}{number of amino acids to retain on each side}
}
\value{
A matrix with extracted kmers
}
\description{
This function extracts simetric k-mers around prolines
}
\note{
This is an internal function used inside predict_hyp
}
\seealso{
\code{\link[ragp]{predict_hyp}}
}
