% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maab.R
\name{maab}
\alias{maab}
\alias{maab.default}
\alias{maab.character}
\alias{maab.data.frame}
\alias{maab.list}
\title{MAAB classification of hydroxyproline rich glycoproteins}
\usage{
maab(data, ...)

\method{maab}{character}(data, ...)

\method{maab}{data.frame}(data, sequence, id, ...)

\method{maab}{list}(data, ...)

\method{maab}{default}(data = NULL, sequence, id, order = c("ext",
  "tyr", "prp", "agp"), gpi = NULL, get_gpi = c("bigpi", "predgpi",
  "netgpi", "none"), spec = 0.99, progress = FALSE, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from \code{\link[seqinr]{read.fasta}} call. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{order}{Order of motif counting, the default is as in Johnson et al. (2017).}

\item{gpi}{A Boolean vector indicating if the corresponding id contains a GPI or not. Can be the 'is.bigpi' column from the output of get_big_pi.}

\item{get_gpi}{A string indicating if \code{\link[ragp]{get_big_pi}},  \code{\link[ragp]{get_pred_gpi}} or \code{\link[ragp]{get_netGPI}} should be called on sequences that belong to one of the HRGP classes thus resolving class ambiguities that depend on GPI knowledge. At default set to 'none'.}

\item{spec}{Numeric in the 0-1 range, indicating the threshold specificity of \code{\link[ragp]{get_pred_gpi}}. Only valid if argument get_gpi = "predgpi".}

\item{progress}{Boolean, whether to show the progress bar, at default set to FALSE.}
}
\value{
A data frame with columns:
\enumerate{
\item{id} {protein identifiers as from input}
\item{ext_sp} {number of extensin SPn motifs, counted using SP{3,5}}
\item{ext_tyr} {number of extensin TYR motifs, sum of matches for: [FY].Y, KHY, VY[HKDE], V.Y, YY}
\item{prp} {number of proline rich protein motifs, sum of matches for: PPV.[KT], PPV[QK], KKPCPP}
\item{agp} {number of arabinogalactan motifs, sum of matches for: [AVTG]P{1,3}, [ASVTG]P{1,2}}
\item{past_percent} {summed percent of "P", "A", "S" and "T" amino acids}
\item{pvyk_percent} {summed percent of "P", "V", "Y" and "K" amino acids}
\item{psky_percent} {summed percent of "P", "S", "K" and "Y" amino acids}
\item{p_percent} {percent of "P"}
\item{coverage} {the coverage of sequence by the identified motifs}
\item{maab_class} {determined maab class}
}
}
\description{
Perform Motif and amino acid bias classification of hydroxyproline rich glycoproteins according to Johnson et al. (2017)
}
\details{
The function provides motif and amino acid bias descriptors used for classification of HRGP's by the MAAB pipeline (Johnson et al. 2017) as well as the determined HRGP classes. The motifs are counted in a specific order ext > tyr > prp > agp, and overlapping motifs are not counted. Hence the classification depends on the order of counting, this is most noticeable for tyr and prp, we recommend using both the default order and  'order = c("ext", "prp","tyr", "agp")'.
}
\examples{

library(ragp)
data(at_nsp)

maab_class <- maab(sequence = at_nsp$sequence,
                   id = at_nsp$Transcript.id)

head(maab_class)

}
\references{
Johnson KL, Cassin AM, Lonsdale A, Bacic A, Doblin MS, Schultz CJ. (2017) Pipeline to Identify Hydroxyproline-Rich Glycoproteins. Plant Physiol 174(2): 886-903.
}
\seealso{
\code{\link[ragp]{scan_ag}}
}
