% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfam2go.R
\name{pfam2go}
\alias{pfam2go}
\title{Add GO terms based on pfam accessions}
\source{
\url{http://geneontology.org/external2go/pfam2go}  

       \url{ftp://ftp.geneontology.org/pub/go/external2go/pfam2go}
}
\usage{
pfam2go(data_pfam, pfam)
}
\arguments{
\item{data_pfam}{a data frame containing a column with PFAM accessions}

\item{pfam}{a string defining the column name where the PFAM accessions are stored. Defaults to "acc" as per output of get_hmm function.}
}
\value{
A merged data frame with columns as provided in data_pfam argument and additional columns:
\describe{
  \item{Pfam_acc}{Character, PFAM family accession.}
  \item{Pfam_name}{Character, PFAM family name.}
  \item{GO_name}{Character, GO term name.}
  \item{GO_acc}{Character, GO term accession.}
  }
}
\description{
The objective of gene ontology (GO) is to provide controlled vocabularies for the description of the biological process, molecular function, and cellular component of gene products. This function maps existing PFAM accessions to corresponding GO terms.
}
\examples{

library(ragp)
data(at_nsp)

pfam_pred <- get_hmm(sequence = at_nsp$sequence[1],
                     id = at_nsp$Transcript.id[1])

pfam_pred_go <- pfam2go(data_pfam = pfam_pred, pfam = "acc")
pfam_pred_go

}
\seealso{
\code{\link[ragp]{get_hmm}}
}
