% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hyp.R
\name{predict_hyp}
\alias{predict_hyp}
\title{Predict hydroxyproline positions in plant proteins based on primary structure}
\usage{
predict_hyp(sequence, id, tprob = 0.33, split = 1)
}
\arguments{
\item{sequence}{A vector of strings representing protein amino acid sequences}

\item{id}{A vector of strings representing protein id's}

\item{tprob}{A numeric value indicating the treshold for prediction. Acceptable values are in 0 - 1 range. At default set to 0.33 which corresponds to 0.9674 sensitivity and 0.9568 specificity in 5-fold crossvalidation.}

\item{split}{A numeric value determining the ratio of vectorized and parallelized computation. Should be left at default, lower to 0 - 1 range if low memory errors occur. Increase at your own risk.}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, indicating the inputed protein id's}
  \item{substr}{Character, indicating the sequence substring which was used for predictions}
  \item{P_pos}{Integer, position of proline in the sequence}
  \item{prob}{Numeric, predicted probability of beeing hydroxyproline}
  }
}
\description{
predict_hyp is a hydroxyproline site prediction algorithm for plant proteins, based on the xgboost distributed gradient boosting library.
It was trained on plant sequences with experimentally determined 4-hydroxyprolines from uniprot data base. N- and C- terminal prolines surrounded by less than 10 amino acid residues will be excluded.
}
\examples{
library(ragp)
data(at_nsp)

#ramdom indexes
ind <- c(129, 145, 147, 160, 170,
    180, 189, 203, 205, 214, 217, 224)

hyp_pred <- predict_hyp(sequence = at_nsp$sequence[ind], id = at_nsp$Transcript.id[ind])
}
