% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_hyp.R
\name{QSOlevel}
\alias{QSOlevel}
\title{Quasi-Sequence-Order (QSO) Descriptor}
\usage{
QSOlevel(m)
}
\arguments{
\item{m}{A vector of k-mers}
}
\value{
A matrix with the Quasi-Sequence-Order Descriptor (dim: 64)
}
\description{
This function calculates the Quasi-Sequence-Order (QSO) descriptor with lag 12
}
\details{
QSO descriptor is calculated based on two physicochemical distance matrices: Schneider-Wrede (Schneider and Wrede, 1994) and Grantham physicochemical distance matrix (Grantham, 1974)
}
\note{
This is an internal function used inside predict_hyp
}
\references{
Kuo-Chen Chou. Prediction of Protein Subcellar Locations by Incorporating Quasi-Sequence-Order Effect. Biochemical and Biophysical Research Communications, 2000, 278, 477-483.
Gisbert Schneider and Paul Wrede. The Rational Design of Amino Acid Sequences by Artifical Neural Networks and Simulated Molecular Evolution: Do Novo Design of an Idealized Leader Cleavage Site. Biophys Journal, 1994, 66, 335-344.
}
\seealso{
\code{\link[ragp]{predict_hyp} \link[protr]{extractQSO}}
}
\author{
original R code by Nan Xiao, modified by Milan Dragićević
}
