% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_nglc.R
\name{scan_nglc}
\alias{scan_nglc}
\title{Detection of motifs for N-glycosylation on asparagine residues.}
\source{
\url{https://prosite.expasy.org/PDOC00001}
}
\usage{
scan_nglc(data = NULL, sequence, id, span = 5L, cutoff = 0, nsp = 15L)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from seqinr::read.fasta call, or an object of class "AAStringSet" resulting from Biostrings::readAAStringSet call.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{span}{An integer specifying how many amino acids around the target asparagine residues is used to calculate hydrophilicity. At default set to 5: asparagine position - 5 to asparagine position +5 residues. Range to consider: 3 - 10. Acceptable values are 0 - 20.}

\item{cutoff}{An integer specifying the cutoff value for hydrophilicity. Range to consider: -1 - 1. Values lower then -3.4 exclude hydrophilicity as a parameter, while values higher than 3 result in no motifs being found.}

\item{nsp}{An integer, either of length 1 or length equal the number of sequences, specifying the number of N-terminal amino acids to exclude.}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, as supplied in the function call.}
  \item{align_start}{Integer, start of motif match.}
  \item{motif}{Character, the motif matched.}
  \item{hydrophilicity}{the average hydrophilicity.}
  \item{is.nglc}{Boolean, is the N-glycosylation likely based on hydrophilicity.}
  \item{nsp}{Optional integer column provided when nsp argument is of equal length to the number of input sequences}
}
}
\description{
Detection is based on PROSITE pattern PS00001. Mean local hydrophilicity (Hopp and Woods, 1981) is used to assess if the asparagines are buried.
}
\note{
For N-glycosylation to happen the protein must enter the endoplasmic reticulum. Please check if the proteins are likely to contain a N-terminal signal peptide. The motif Asp-Xaa-Ser/Thr (where Xaa is not Pro) on which N-glycosylation occurs is relatively common, however for N-glycosylation to occur the motif needs to be located on the protein surface. Mean local hydrophilicity (Hopp and Woods, 1981) is used here to evaluate if the asparagines are located in a hydrophilic surrounding which is more likely on the protein surface.
}
\examples{
library(ragp)
data(at_nsp)

nglc_pred <- scan_nglc(data = at_nsp,
                       sequence = sequence,
                       id = Transcript.id)

}
\references{
Hopp TP. Woods KR. (1981) Prediction of protein antigenic determinants from amino acid sequences. Proceedings of the National Academy of Sciences of the United States of America, 78(6): 3824-8
}
\author{
original R code by Thomas Shafee, modified by Milan Dragićević
}
