% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_hyp.R
\name{extractMBdesc}
\alias{extractMBdesc}
\title{Normalized Moreau-Broto Autocorrelation Descriptor}
\usage{
extractMBdesc(x)
}
\arguments{
\item{m}{A vector of k-mers}
}
\value{
A matrix with the Normalized Moreau-Broto Autocorrelation Descriptor (dim: 36)
}
\description{
This function calculates the normalized Moreau-Broto autocorrelation descriptor with lag 6.
}
\details{
Normalized Moreau-Broto Autocorrelation Descriptor is calculated based on the following amino acid attributes:
CIDH920105 - Normalized average hydrophobicity scales, BHAR880101 - Average flexibility indices,
CHAM820102 - Free energy of solution in water, kcal/mole, BIGC670101 - Residue volume,
CHAM810101 - Polarizability parameter and DAYM780201 - Relative mutability
}
\note{
This is an internal function used inside predict_hyp
}
\seealso{
\code{\link[ragp]{predict_hyp} \link[protr]{extractMoreauBroto}}
}
