% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_big_pi.R
\name{get_big_pi}
\alias{get_big_pi}
\title{Scraping big-PI Plant Predictor web server.}
\source{
\url{http://mendel.imp.ac.at/gpi/plant_server.html}
}
\usage{
get_big_pi(sequence, short.out = T, sleep = NULL, verbose = NULL)
}
\arguments{
\item{sequence}{A vector of strings representing protein amino acid sequences}

\item{short.out}{A bolean indicating the type of returned object, defaults to T}

\item{sleep}{A numeric indicating the pause in seconds betwean server calls, at default set to 2}

\item{verbose}{Bolean wheather to print out the total score for each sequence, defaults to T}
}
\value{
If short.out==T:
A data frame with columns:
\describe{
  \item{quality_gpi}{Character, indicating the "Quality of the highest scoring omega-site"}
  \item{position_gpi}{Iinteger, indicating the "Sequence position of the highest scoring omega-site"}
  \item{total_score}{Numeric, indicating "Total Score of the highest scoring omega-site"}
  \item{profile_score}{Numeric, indicating the "Profile Score highest scoring omega-site"}
  \item{profile_i_score}{Numeric, indicating "Profile independent Score of the highest scoring omega-site"}
  \item{alt_quality_gpi}{Character, indicating the "Quality of the alternative omega-site", if present, else NA}
  \item{alt_position_gpi}{Integer, indicating the "Sequence position of the alternative omega-site", if present, else NA}
  \item{alt_total_score}{Numeric, indicating "Total Score of the alternative omega-site", if present, else NA}
  \item{alt_profile_score}{Numeric, indicating the "Profile Score of the alternative omega-site", if present, else NA}
  \item{alt_profile_i_score}{Numeric, indicating "Profile independent Score of the alternative omega-site" if present, else NA}
  }

If short.out==F:
A list where each element correspondes to one queried sequence containing lines of text of the whole big-PI Plant Predictor output.
}
\description{
big-PI Plant Predictor is a web server utlizing a scoring algorithm for prediction of GPI modification sites in plants.
}
\examples{
library(ragp)
data(at_nsp)

#indexes of some sequences known to have a positive prediction
ind <- c(129, 145, 147, 160, 170,
    180, 189, 203, 205, 214, 217, 224)

big_pi_pred <- get_big_pi(sequence = at_nsp$sequence[ind])

#extract terms not included in the short output:

#first define a function for extraction

extract_val <- function(x, y){
 res <- regmatches(y[grep(x, y)],
             gregexpr(num_reg, y[grep(x, y)]))
 res <- unlist(res)
 res <- as.numeric(res)
 return(res)
}

#define a regex that extracts numerics
num_reg <- "[-+]?[0-9]*\\\\.?[0-9]+([eE][-+]?[0-9]+)?"

#run function with short.out=F
big_pi_pred <- get_big_pi(sequence = at_nsp$sequence[ind],
                          short.out = F)

lapply(big_pi_pred,
      function(z) extract_val(x="Term  3", y=z)[4])

lapply(test_big_po,
      function(z) extract_val(x = "Total Score", y = z)[1])

#interested in the alternative site:
lapply(big_pi_pred,
      function(x) extract_val(x = "Term  3", y = x)[5])
#NAs do not have the alternative site

}
\references{
Eisenhaber B. Wildpaner M. Schultz CJ. Borner GHH. Dupree P. Eisenhaber F. (2003) Glycosylphosphatidylinositol lipid anchoring of plant proteins. Sensitive prediction from sequence- and genome-wide studies for Arabidopsis and rice. Plant Physiology 133(4): 1691-701
}
