% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hmm.R
\name{get_hmm}
\alias{get_hmm}
\alias{get_hmm.default}
\alias{get_hmm.character}
\alias{get_hmm.data.frame}
\alias{get_hmm.list}
\alias{get_hmm.AAStringSet}
\title{Query hmmer web server.}
\source{
\url{https://www.ebi.ac.uk/Tools/hmmer/search/hmmscan}
}
\usage{
get_hmm(data, ...)

\method{get_hmm}{character}(data, ...)

\method{get_hmm}{data.frame}(data, sequence, id, ...)

\method{get_hmm}{list}(data, ...)

\method{get_hmm}{default}(
  data = NULL,
  sequence,
  id,
  verbose = FALSE,
  sleep = 1,
  attempts = 2L,
  timeout = 10,
  progress = FALSE,
  ievalue = NULL,
  bitscore = NULL,
  ...
)

\method{get_hmm}{AAStringSet}(data, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from \code{\link[seqinr]{read.fasta}} call. Alternatively an `AAStringSet` object. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{verbose}{Boolean, whether to print out the output for each sequence, defaults to FALSE.}

\item{sleep}{Numeric indicating the pause in seconds between server calls, at default set to 1.}

\item{attempts}{Integer, number of attempts if server unresponsive, at default set to 2.}

\item{timeout}{Numeric, time in seconds to wait for server response.}

\item{progress}{Boolean, whether to show the progress bar, at default set to FALSE.}

\item{ievalue}{Numeric, all sequences with independent E-value lower or equal to this value will be retained in the function output. Used to filter out low similarity matches. If set some queried sequences might be discarded from the output. Suggested values: 1e-2 - 1e-5.}

\item{bitscore}{Numeric, all sequences with bitscore greater or equal to this value will be retained in the function output. Used to filter out low similarity. If set some queried sequences might be discarded from the output. Suggested values: 10 - 20.}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, as supplied in the function call}
  \item{name}{Character, PFAM family name}
  \item{acc}{Character, PFAM family accession}
  \item{desc}{Character, PFAM family description}
  \item{clan}{Character, PFAM clan}
  \item{align_start}{Numeric, start of domain alignment in query sequence}
  \item{align_end}{Numeric, end of domain alignment in query sequence}
  \item{model_start}{Numeric, start of alignment in domain model}
  \item{model_end}{Numeric, end of alignment in domain model}
  \item{ievalue}{Numeric, the "independent E-value", the E-value that the sequence/profile comparison would have received if this were the only domain envelope found in it, excluding any others. This is a stringent measure of how reliable this particular domain may be. The independent E-value uses the total number of targets in the target database.}
  \item{cevalue}{Numeric, the "conditional E-value", a permissive measure of how reliable this particular domain may be.}
  \item{bitscore}{Numeric, the domain bit score.}
  \item{reported}{Logical, is the result reported on the hmmer site. The hmmer web server outputs more hmm profile matches than it presents to the user. Results below a certain threshold are not reported (hidden) on the site.}
}
}
\description{
hmmer web server offers biosequence analysis using profile hidden Markov Models. This function allows searching
of a protein sequence vs a profile-HMM database (Pfam-A).
}
\note{
hmmscan does not handle sequences longer than 1000 amino acids. get_hmm splits these sequences into shorter substrings which overlap by 300 amino acids and queries hmmscan. Some results might be redundant or partially overlapping in this case. When this is an issue it is advisable to provide a subsequence of appropriate length as get_hmm input.
}
\examples{

pfam_pred <- get_hmm(data = at_nsp[1:5,],
                    sequence = sequence,
                    id = Transcript.id,
                    verbose = FALSE,
                    sleep = 0)
pfam_pred

}
\seealso{
\code{\link[ragp]{pfam2go}}
}
