% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hmm.R
\name{get_hmm}
\alias{get_hmm}
\title{Scraping hmmer web server.}
\source{
\url{https://www.ebi.ac.uk/Tools/hmmer/search/hmmscan}
}
\usage{
get_hmm(sequence, id, verbose = NULL, sleep = NULL)
}
\arguments{
\item{sequence}{A vector of strings representing protein amino acid sequences}

\item{id}{A vector of strings representing the names of the corresponding sequences}

\item{verbose}{Bolean wheather to print out the output for each sequence, defaults to T}

\item{sleep}{Numeric indicating the pause in seconds betwean server calls, at default set to 1}
}
\value{
A data frame with columns:
\describe{
  \item{id}{Character, as suplied in the function call}
  \item{name}{Character, PFAM family name}
  \item{acc}{Character, PFAM family accession}
  \item{desc}{Character, PFAM family description}
  \item{clan}{Character, PFAM clan}
  \item{align_start}{Numeric, start of domain alignment in query sequence}
  \item{align_end}{Numeric, end of domain alignment in query sequence}
  \item{model_start}{Numeric, start of alignment in domain model}
  \item{model_end}{Numeric, end of alignment in domain model}
  \item{ievalue}{Numeric, the "independent E-value", the E-value that the sequence/profile comparison would have received if this were the only domain envelope found in it, excluding any others. This is a stringent measure of how reliable this particular domain may be. The independent E-value uses the total number of targets in the target database.}
  \item{cevalue}{Numeric, the "conditional E-value", a permissive measure of how reliable this particular domain may be.}
  \item{bitscore}{Numeric, the domain bit score.}
  \item{reported}{Logical, is the result reported on the hmmer site. The hmmer web server outputs more hmm profile matches than it presents to the user. Results below a certain treshold are not reported (hidden) on the site.}
}
}
\description{
hmmer web server offers biosequence analysis using profile hidden Markov Models. This function allows searching
of a protein sequence vs a profile-HMM database (Pfam-A).
}
\examples{

pfam_pred = get_hmm(sequence = at_nsp$sequence[1:20],
                  id = at_nsp$Transcript.id[1:20])

}
\seealso{
\code{\link[ragp]{pfam2go}}
}
