% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred_gpi.R
\name{get_pred_gpi}
\alias{get_pred_gpi}
\alias{get_pred_gpi.default}
\alias{get_pred_gpi.character}
\alias{get_pred_gpi.data.frame}
\alias{get_pred_gpi.list}
\alias{get_pred_gpi.character}
\alias{get_pred_gpi.data.frame}
\alias{get_pred_gpi.list}
\alias{get_pred_gpi.default}
\title{Query PredGPI web server.}
\source{
\url{http://gpcr.biocomp.unibo.it/predgpi/pred.htm}
}
\usage{
get_pred_gpi(data, ...)

\method{get_pred_gpi}{character}(data, spec = 0.99, ...)

\method{get_pred_gpi}{data.frame}(data, sequence, id, ...)

\method{get_pred_gpi}{list}(data, ...)

\method{get_pred_gpi}{default}(data = NULL, sequence, id, ...)
}
\arguments{
\item{data}{A data frame with protein amino acid sequences as strings in one column and corresponding id's in another. Alternatively a path to a .fasta file with protein sequences. Alternatively a list with elements of class "SeqFastaAA" resulting from \code{\link[seqinr]{read.fasta}} call. Should be left blank if vectors are provided to sequence and id arguments.}

\item{...}{currently no additional arguments are accepted apart the ones documented bellow.}

\item{spec}{Numeric in the 0-1 range, indicating the threshold specificity.}

\item{sequence}{A vector of strings representing protein amino acid sequences, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}

\item{id}{A vector of strings representing protein identifiers, or the appropriate column name if a data.frame is supplied to data argument. If .fasta file path, or list with elements of class "SeqFastaAA" provided to data, this should be left blank.}
}
\value{
A data frame with columns:
\describe{
\item{id}{Character, name of the submitted sequence.}
\item{omega_site}{Integer, indicating the sequence position of the omega-site.}
\item{specificity}{Numeric, indicating the specificity of the prediction.}
\item{is.gpi}{Logical, is the specificity of the prediction higher than the set threshold specificity}
}
}
\description{
PredGPI web server is a predictor of GPI modification sites.
}
\note{
This function creates temporary files in the working directory.
}
\examples{
library(ragp)
data(at_nsp)

gpi_pred <- get_pred_gpi(at_nsp[1:20,],
                         sequence,
                         Transcript.id)

}
\references{
Pierleoni A. Martelli P. L. Casadio R. (2008) PredGPI: a GPI-anchor predictor. BMC Bioinformatics 9: 392
}
\seealso{
\code{\link[ragp]{get_big_pi}}
}
