% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maab.R
\name{maab}
\alias{maab}
\title{Motif and amino acid bias of a protein sequence}
\usage{
maab(sequence, id, frame = NULL, custom = NULL)
}
\arguments{
\item{sequence}{A vector of strings representing protein amino acid sequences}

\item{id}{A vector of strings representing the names of the corresponding sequences}

\item{frame}{An optional integer defining the frame length for sliding along the sequences}

\item{custom}{An optional character vector which can contain amino acid one letter symbols eg. c("P", "V", "K", "C", "T"), strings corresponding to amino acid motifs eg c("PTYK", "PVKC"), mixing of types is supported. Common regex operators can also be utilized.}
}
\value{
by default if, frame is not specified a data frame is returned with columns:
\enumerate{
\item{past_count:} {summed count of "P", "A", "S" and "T" amino acids}
\item{pvyk_count:} {summed count of "P", "V", "Y" and "K" amino acids}
\item{psky_count:} {summed count of "P", "S", "K" and "Y" amino acids}
\item{p_count:} {count of "P"}
\item{ext_sp_count:} {summed count of SPPP+ motifs}
\item{ext_fyxy_count:} {summed count of [FY]XY motifs}
\item{ext_khy_count:} {summed count of KHY motifs}
\item{ext_vyhkde_count:} {summed count of VY[HKDE] motifs}
\item{ext_vxy_count:} {summed count of VXY motifs}
\item{ext_yy_count:} {summed count of YY motifs}
\item{prp_ppvqk_count:} {summed count of PPV[QK] motifs}
\item{prp_ppvxkt_count:} {summed count of PPVX[KT] motifs}
\item{ppr_kkpcpp:} {summed count of KKPCPP motifs}
\item{past_percent:} {summed percent of "P", "A", "S" and "T" amino acids}
\item{pvyk_percent:} {summed percent of "P", "V", "Y" and "K" amino acids}
\item{psky_percent:} {summed percent of "P", "S", "K" and "Y" amino acids}
\item{p_percent:} {percent of "P"}
\item{id:} {names of the corresponding sequences, as per input}
}

if frame is specified a list with two elements is returned:
\enumerate{
\item{protein:} {a named list with names corresponding to sequence id's, where each element is a data frame resembling the one returned when frame = NULL, each row of which corresponds to each sequence frame}
\item{maab:} {a character vector with search criteria}
}

if custom is specified the returned object contains counts for each element specified in custom; additionally percentage is calculated for all elements containing a single symbol.
}
\description{
Calculate the compositional bias of whole protein sequences, or slide along the sequences and calculate the bias for all subsequences of a defined length (frame). Sequence motifs can also be searched.
}
\details{
By default the function provides motif and amino acid bias descriptors used for classification of HRGP's by the MAAB pipeline (Johnson et al. 2017). Arabinogalactan descriptors used in the MAAB pipeline other than the percent of PAST amino acids are not incorporated in this function. Users can specify a sliding frame window in order to calculate partial sequence descriptors or specify custom search criteria.
}
\references{
Johnson KL, Cassin AM, Lonsdale A, Bacic A, Doblin MS, Schultz CJ. (2017) Pipeline to Identify Hydroxyproline-Rich Glycoproteins. Plant Physiol 174(2): 886-903.
}
\seealso{
\code{\link[ragp]{scan_ag}}
}
