% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prot.R
\name{plot_prot}
\alias{plot_prot}
\title{Protein structure diagram.}
\usage{
plot_prot(sequence, id, hyp_col = "#868686FF", gpi_col = "#0073C2FF",
  nsp_col = "#CD534CFF", ag_col = "#E5E5E5FF", tm_col = "#EFC000FF",
  hyp = TRUE, gpi = c("bigpi", "predgpi", "none"), nsp = TRUE,
  ag = TRUE, tm = TRUE, domain = TRUE, disorder = FALSE,
  dom_sort = c("ievalue", "abc", "cba"), progress = FALSE, ...)
}
\arguments{
\item{sequence}{String representing a protein amino acid sequence.}

\item{id}{String representing a protein identifier. Will be converted using \code{\link[base]{make.names}}}

\item{hyp_col}{Plotting color of predicted hydroxyproline positions. At default set to: '#868686FF'.}

\item{gpi_col}{Plotting color of the predicted omega site (glycosylphosphatidylinositol attachment). At default set to: '#0073C2FF'.}

\item{nsp_col}{Plotting color of the N-terminal signal peptide. At default set to: '#CD534CFF'.}

\item{ag_col}{Plotting color of the AG glycomodul spans. At default set to: '#E5E5E5FF'.}

\item{tm_col}{Plotting color of the transmembrane regions. At default set to: '#EFC000FF'.}

\item{hyp}{Bolean, should hydroxyprolines be plotted.}

\item{gpi}{A string indicating if \code{\link[ragp]{get_big_pi}} (gpi = "bigpi") or \code{\link[ragp]{get_pred_gpi}} (gpi = "predgpi") should be called when predicting omega sites. To turn off omega site prediction use gpi = "none". At default set to "bigpi".}

\item{nsp}{Bolean, should the N-terminal signal peptide be plotted.}

\item{ag}{Bolean, should the AG glycomodul spans be plotted.}

\item{tm}{Bolean, should the transmembrane regions be plotted.}

\item{domain}{Bolean, should the domains be plotted.}

\item{disorder}{Bolean, should disordered regions be plotted.}

\item{dom_sort}{One of c("ievalue", "abc", "cba"), defaults to "abc". Domain plotting order. If 'ievalue' domains with the lowest ievalue as determined by hmmscan will be plotted above. If 'abc' or 'cba' the order is determined by domain Names.}

\item{progress}{Bolean, whether to show the progress bar, at default set to FALSE.}

\item{...}{Appropriate arguments passed to \code{\link[ragp]{get_signalp}}, \code{\link[ragp]{get_espritz}}, \code{\link[ragp]{predict_hyp}} and \code{\link[ragp]{scan_ag}}.}
}
\value{
A ggplot2 plot object
}
\description{
Plots a diagram of protein structure based on hmmscan domain annotation and several types of predictions.
}
\examples{
library(ragp)
library(ggplot2)
ind <- c(23, 24, 5, 80, 81, 120, 230, 334, 345, 1000)
pred <- plot_prot(sequence = at_nsp$sequence[ind],
                  id = at_nsp$Transcript.id[ind])
pred +
  theme(legend.position = "bottom",
        legend.direction = "vertical")

}
\seealso{
\code{\link[ragp]{get_signalp}} \code{\link[ragp]{get_phobius}} \code{\link[ragp]{get_hmm}} \code{\link[ragp]{get_espritz}} \code{\link[ragp]{predict_hyp}} \code{\link[ragp]{scan_ag}}
}
