% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hyp.R
\name{predict_hyp}
\alias{predict_hyp}
\title{Predict hydroxyproline positions in plant proteins based on primary structure}
\usage{
predict_hyp(sequence, id, tprob = 0.32, split = 1)
}
\arguments{
\item{sequence}{A vector of strings representing protein amino acid sequences}

\item{id}{A vector of strings representing protein id's}

\item{tprob}{A numeric value indicating the treshold for prediction. Acceptable values are in 0 - 1 range. At default set to 0.32 offering a tradeoff between sensitivity and specificity.}

\item{split}{A numeric value determining the ratio of vectorized and parallelized computation. Should be left at default, lower to 0 - 1 range if low memory errors occur. Increase at your own risk.}
}
\value{
A list with two elements:
\describe{
  \item{prediction}{data frame with columns:
  id - character, indicating the inputed protein id's;
  substr - character, indicating the sequence substring which was used for predictions;
  P_pos - integer, position of proline in the sequence;
  prob - numeric, predicted probability of beeing hydroxyproline}
  \item{sequence}{sequences with prolines - P substituted with hydroxyprolines - O according to the prediction}
}
}
\description{
predict_hyp is a hydroxyproline site prediction algorithm for plant proteins, based on the xgboost distributed gradient boosting library.
It was trained on plant sequences with experimentally determined 4-hydroxyprolines from uniprot data base. N- and C- terminal prolines surrounded by less than 10 amino acid residues will be excluded.
}
\examples{
library(ragp)
data(at_nsp)

#ramdom indexes
ind <- c(129, 145, 147, 160, 170,
    180, 189, 203, 205, 214, 217, 224)

hyp_pred <- predict_hyp(sequence = at_nsp$sequence[ind], id = at_nsp$Transcript.id[ind])
}
